there's now a fully functional login page
except login is still not enforced and almost every request is still
handled via my own API key configured in .env file. BUT it does check
for valid r/w API key and it does flash the user with a personal welcome
pop-up message.
Next in line would be to store user fullname and API key in cookie then
check for saved credential and if not found redirect to login page.
Also, logout button to clear session.